+++ core/src/kriuchkov/maksim/spaceshooter/SpaceShooter.java
@@ -1,36 +1,14 @@
 package kriuchkov.maksim.spaceshooter;
 
-import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.Game;
 
-public class SpaceShooter extends ApplicationAdapter {
-	SpriteBatch batch;
-	Texture img;
-	Texture background;
-	
-	@Override
-	public void create () {
-		batch = new SpriteBatch();
-		img = new Texture("badlogic.jpg");
-		background = new Texture("background_simple.png");
-	}
+import kriuchkov.maksim.spaceshooter.screen.MenuScreen;
 
-	@Override
-	public void render () {
-		Gdx.gl.glClearColor(1, 0, 0, 1);
-		Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-		batch.begin();
-		batch.draw(background, 0, 0);
-		batch.draw(img, 0, 0);
-		batch.end();
-	}
-	
-	@Override
-	public void dispose () {
-		batch.dispose();
-		img.dispose();
-	}
+public class SpaceShooter extends Game {
+
+    @Override
+    public void create () {
+        setScreen(new MenuScreen());
+    }
+
 }
